flask/requirements.txt
python-dotenv
flask
waitress
flask-cors
psycopg2